requirements.txt
django
django-admin-oauth2
django-environ
python-dotenv
djangorestframework
drf_spectacular
django-simple-history
django-debug-toolbar
django-bootstrap5
psycopg2-binary

libsass
django-compressor
django-sass-processor

yandex_music

undetected-chromedriver
pandas
scikit-learn
